Add chain id checks

+++ examples/cross-chain-warriors/contracts/test/CrossChainWarriorsZetaMPIMock.sol
@@ -8,7 +8,7 @@ import "../CrossChainWarriors.sol";
 contract CrossChainWarriorsZetaMPIMock is ZetaMPI {
     function callOnZetaMessage(
         bytes memory originSenderAddress,
-        uint16 originChainId,
+        uint256 originChainId,
         address destinationAddress,
         uint256 zetaAmount,
         bytes calldata message
@@ -27,7 +27,7 @@ contract CrossChainWarriorsZetaMPIMock is ZetaMPI {
 
     function callOnZetaRevert(
         address originSenderAddress,
-        uint16 destinationChainId,
+        uint256 destinationChainId,
         bytes calldata destinationAddress,
         uint256 zetaAmount,
         uint256, // gasLimit
@@ -46,7 +46,7 @@ contract CrossChainWarriorsZetaMPIMock is ZetaMPI {
     }
 
     function send(ZetaInterfaces.SendInput calldata _input) external {
-        uint16 originChainId = _input.destinationChainId == 2 ? 1 : 2;
+        uint256 originChainId = _input.destinationChainId == 2 ? 1 : 2;
         address dest = abi.decode(_input.destinationAddress, (address));
 
         return callOnZetaMessage(abi.encode(msg.sender), originChainId, dest, _input.zetaAmount, _input.message);
